refactor(core): Ollama-Typen und interne Interfaces aus Core herauslösen
- IOllamaClient, IAiSettingsService, OllamaSettings, OllamaModelInfo, OllamaStatus
  nach ghGPT.Ai.Ollama verschoben — Core bleibt technologie-agnostisch
- IToolDispatcher nach ghGPT.Ai.Tools verschoben (Implementierung liegt dort)
- IChatContextBuilder als internal nach ghGPT.Ai verschoben (nur intern genutzt)
- ChatHistoryEntry in eigene Datei extrahiert (eine Klasse pro Datei)
- ghGPT.Api erhält direkte Referenz auf ghGPT.Ai.Ollama
