+++ src/ghGPT.Ai.Tools/IToolDispatcher.cs
@@ -1,4 +1,6 @@
-namespace ghGPT.Core.Ai;
+using ghGPT.Core.Ai;
+
+namespace ghGPT.Ai.Tools;
 
 public interface IToolDispatcher
 {
