+++ src/ghGPT.Ai/ChatService.cs
@@ -1,3 +1,4 @@
+using ghGPT.Ai.Ollama;
 using ghGPT.Ai.Tools;
 using ghGPT.Core.Ai;
 using System.Runtime.CompilerServices;
