+++ src/ghGPT.Ai/IChatContextBuilder.cs
@@ -1,0 +1,8 @@
+using ghGPT.Core.Ai;
+
+namespace ghGPT.Ai;
+
+internal interface IChatContextBuilder
+{
+    Task<IEnumerable<ChatMessage>> BuildAsync(ChatRequest request);
+}
