+++ src/ghGPT.Api/ghGPT.Api.csproj
@@ -13,6 +13,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\ghGPT.Core\ghGPT.Core.csproj" />
+    <ProjectReference Include="..\ghGPT.Ai.Ollama\ghGPT.Ai.Ollama.csproj" />
     <ProjectReference Include="..\ghGPT.Infrastructure\ghGPT.Infrastructure.csproj" />
   </ItemGroup>
