+++ src/ghGPT.Core/Ai/ChatHistoryEntry.cs
@@ -1,0 +1,8 @@
+namespace ghGPT.Core.Ai;
+
+public class ChatHistoryEntry
+{
+    public string Role { get; set; } = string.Empty;
+    public string Content { get; set; } = string.Empty;
+    public DateTimeOffset Timestamp { get; set; }
+}
