--- src/ghGPT.Core/Ai/IChatContextBuilder.cs
@@ -1,6 +1,0 @@
-namespace ghGPT.Core.Ai;
-
-public interface IChatContextBuilder
-{
-    Task<IEnumerable<ChatMessage>> BuildAsync(ChatRequest request);
-}
