+++ src/ghGPT.Core/Ai/IChatHistoryService.cs
@@ -6,10 +6,3 @@ public interface IChatHistoryService
     void Append(string repoId, string role, string content);
     void Clear(string repoId);
 }
-
-public class ChatHistoryEntry
-{
-    public string Role { get; set; } = string.Empty;
-    public string Content { get; set; } = string.Empty;
-    public DateTimeOffset Timestamp { get; set; }
-}
